# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradersHub professional trading platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # WALLET
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get all trades for the current user and selected account
      parameters:
        - name: account
          in: query
          schema:
            type: string
            enum: [real, demo]
            default: real
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Open a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Validation error or insufficient balance
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [wallet]
      summary: Get deposit history for the selected account
      parameters:
        - name: account
          in: query
          schema:
            type: string
            enum: [real, demo]
            default: real
      responses:
        "200":
          description: Deposit history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"

  # DEMO ACCOUNT
...
  /admin/users/{id}/balance:
    patch:
      operationId: adjustBalance
      tags: [admin]
      summary: Adjust user balance (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceAdjustInput"
      responses:
        "200":
          description: Balance adjusted
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        userId:
          type: integer
        requiresVerification:
          type: boolean
        emailSent:
          type: boolean
        devOtp:
          type: string
        message:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
        isAdmin:
          type: boolean
...
    Trade:
      type: object
      required: [id, userId, symbol, direction, amount, entryPrice, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountType:
          type: string
          enum: [real, demo]
        ticketNumber:
          type: string
          nullable: true
        symbol:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        amount:
          type: number
        lotSize:
          type: number
          nullable: true
        marginUsed:
          type: number
          nullable: true
        entryPrice:
          type: number
        closePrice:
          type: number
          nullable: true
        stopLoss:
          type: number
          nullable: true
        takeProfit:
          type: number
          nullable: true
        profitLoss:
          type: number
          nullable: true
        profitLossPercent:
          type: number
          nullable: true
        status:
          type: string
          enum: [open, closed]
        isCopied:
          type: boolean
        copiedFromUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
        closedAt:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
    TradeInput:
      type: object
      required: [symbol, direction, amount]
      properties:
        symbol:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        amount:
          type: number
        lotSize:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        accountType:
          type: string
          enum: [real, demo]
          default: real
    Deposit:
      type: object
      required: [id, userId, amount, status, phone, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountType:
          type: string
          enum: [real, demo]
        amount:
          type: number
        status:
          type: string
        phone:
          type: string
        mpesaReceiptNumber:
          type: string
          nullable: true
        checkoutRequestId:
          type: string
          nullable: true
        createdAt:
          type: string
        userName:
          type: string
          nullable: true
...
    BalanceAdjustInput:
      type: object
      required: [balance]
      properties:
        balance:
          type: number
        reason:
          type: string
...